[FIX] helpers: fix studio xmlids origins bug in `get_ids()`
When a partial studio xmlid is passed to `get_ids()` that matches the `STUDIO_XMLID_RE` regex it will be expanded to a full xmlid. With `mapped=True`, when that happens, the expanded xmlid is returned as origin of the resolved id instead of the source one. This in turn makes `views.edit_views()` fail at L381, and possibly at L394 (if all passed xmlids were expanded).

Changes:
- separated expand studio xmlid to its own helper function `expand_studio_xmlids()`
- add a mapping of xmlids origins inside `get_ids()` so that we keep track of xmlids that got expanded (N.B. both origin and expanded will be returned with `mapped=True`, so 2 results from one xmlid, although with the same mapped ids)
- log a warning in `views.edit_views()` if nothing was done, and skip `arch_update` query if nothing to pass to it
- don't return a `defaultdict` from `get_ids()` but a normal one

closes odoo-ps/custom-util-archive#10