[IMP] operations: del attribute is not idempotent
The `XPathOperation(UpdateAttributes)` raises an error if it does not find the attribute to delete when `None` is passed as the attribute value. This means that repeated test runs fail unless a fresh db is used.

Other operations only raise a `_logger.warning()` if the xpath does not yield any results. Errors are only raised for bad syntax, etc. This commit makes the behaviour more consistent.

closes odoo-ps/custom-util-archive#18