[FIX] views.edit: Fix `NoneType` error with `set_studio_view`
When using the `set_studio_view` method for an upgrade to version 18.0,
Odoo raises a `NoneType` error which checks the attribute `server_port`.

The traceback, when using `set_studio_view`, was the following:
```
  File "/home/odoo/odev/virtualenvs/18.0/lib/python3.10/site-packages/odoo/upgrade/custom_util/views/edit.py", line 401, in set_studio_view
    with mute_logger("odoo.tests.common"), MockRequest(env, context=dict(studio=True)):
  File "/usr/lib/python3.10/contextlib.py", line 135, in __enter__
    return next(self.gen)
  File "/home/odoo/odev/worktrees/18.0/odoo/addons/website/tools.py", line 30, in MockRequest
    if HttpCase.http_port():
  File "/home/odoo/odev/worktrees/18.0/odoo/odoo/tests/common.py", line 1909, in http_port
    return odoo.service.server.server.httpd.server_port
AttributeError: 'NoneType' object has no attribute 'server_port'
```

To avoid this issue, a fix was found and used already by multiple people
directly in the scripts for customers. This commit adds the found fix to
avoid retyping it for each customer who wants to upgrade to version 18.0.